Change default behavior to raise for extra params
Catching potential problems early is really important for test driven
workflows and for rapid development cycles. By only logging in test and
development environments when un-permitted parameters are passed into the
controller, we open the door for confusion when a new parameter is
intentionally introduced but is quietly removed by strong_parameters.

This changes the default, unspecified behavior from logging extra
parameters to raising errors for them.